+++ types/mas.sh
@@ -15,7 +15,12 @@ case $action in
         baking_platform_is "Darwin" || return $STATUS_UNSUPPORTED_PLATFORM
         needs_exec "mas" || return $STATUS_FAILED_PRECONDITION
         bake mas account
-        [ "$?" -gt 0 ] && return $STATUS_FAILED_PRECONDITION
+        # Workaround for mas account failing on macOS >= 12 (i.e. Darwin >= 21)
+        # https://github.com/borksh/bork/issues/42
+        if [ "$?" -gt 0 ]; then
+          release=$(get_baking_platform_release)
+          str_matches "$release" "^21\." || return $STATUS_FAILED_PRECONDITION
+        fi
         bake mas list | grep -E "^$appid" > /dev/null
         [ "$?" -gt 0 ] && return $STATUS_MISSING
         bake mas outdated | grep -E "^$appid" > /dev/null